+++ src/main/java/insane96mcp/iguanatweaksreborn/module/misc/Nerfs.java
@@ -41,9 +41,6 @@ public class Nerfs extends Feature {
 	@Label(name = "No Coordinates", description = "If true, renderDebugInfo is enabled by default. Requires a world restart. This also changes Caverns and Chasms from displaying coordinates with compasses.")
 	public static Boolean noCoordinates = true;
 	@Config
-	@Label(name = "Reduced Random Tick Speed", description = "If true, randomTickSpeed is set to 2 from 3")
-	public static Boolean reducedRandomTickSpeed = true;
-	@Config
 	@Label(name = "Reduced mob cramming", description = "If true, maxEntityCramming game rule is set to 6 from 24")
 	public static Boolean reducedMobCramming = true;
 	@Config
@@ -139,8 +136,6 @@ public class Nerfs extends Feature {
 
 		if (noCoordinates && FMLLoader.isProduction())
 			event.getServer().getGameRules().getRule(GameRules.RULE_REDUCEDDEBUGINFO).set(true, event.getServer());
-		if (reducedRandomTickSpeed)
-			event.getServer().getGameRules().getRule(GameRules.RULE_RANDOMTICKING).set(2, event.getServer());
 		if (reducedMobCramming)
 			event.getServer().getGameRules().getRule(GameRules.RULE_MAX_ENTITY_CRAMMING).set(6, event.getServer());
 	}